+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased](https://github.com/lumik/spycor/compare/v0.1.0...develop)
 
 ### Added
+- [Enhancement #2](https://github.com/lumik/spycor/issues/2): Enabled different settings for individual spectra.
 
 ### Changed
 - `around:` renamed to `extra:`
+- [Wantfix #5](https://github.com/lumik/spycor/issues/5): Removed false spike detection hook.
 
 ### Fixed
 - [Bug #7](https://github.com/lumik/spycor/issues/7): y axis didn't rescale in new Matlab.